+++ test/libsolidity/syntaxTests/denominations/gwei.sol
@@ -1,3 +1,0 @@
-contract C {
-	uint constant gwei = 1 gwei;
-}