+++ Changelog.md
@@ -5,6 +5,7 @@ Breaking changes:
  * Constructors should not have visibility.
  * Deprecated dot syntax for `value` and `gas`.
  * Deprecated the identifier `now`.
+ * Disallow `gwei` as identifier.
  * JSON AST: Removes members with ``null`` value from JSON output.
  * Parser: NatSpec comments on variables are only allowed for public state variables.
  * Type Checker: Disallow shifts by signed types.